fix(plot): separate ColorSelector
