+++ src/caret_analyze/plot/bokeh/plot_util.py
@@ -12,7 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import List, Tuple
+from typing import Dict, List, Sequence, Tuple
+
+from bokeh.colors import Color, RGB
+
+import colorcet as cc
 
 import pandas as pd
 
@@ -92,3 +96,29 @@ def get_freq_with_timestamp(
             diff_base = int(diff*10**(-9))
 
     return pd.Series(timestamp_list), pd.Series(frequency_list)
+
+
+class PlotColorSelector:
+
+    def __init__(self) -> None:
+        self._palette: Sequence[Color] = \
+            [self._from_rgb(*rgb) for rgb in cc.glasbey_bw_minc_20]
+        self._color_map: Dict[str, Color] = {}
+
+    def get_color(
+        self,
+        plot_object_name: str
+    ) -> Color:
+        if plot_object_name not in self._color_map:
+            color_index = len(self._color_map) % len(self._palette)
+            self._color_map[plot_object_name] = self._palette[color_index]
+
+        return self._color_map[plot_object_name]
+
+    @staticmethod
+    def _from_rgb(r: float, g: float, b: float) -> RGB:
+        r_ = int(r*255)
+        g_ = int(g*255)
+        b_ = int(b*255)
+
+        return RGB(r_, g_, b_)
