+++ src/caret_analyze/plot/bokeh/util.py
@@ -12,15 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import abstractmethod
 from logging import getLogger
-from typing import Dict, Optional, Sequence, Tuple, Union
+from typing import Sequence, Tuple, Union
 
-from bokeh.colors import Color, RGB
 from bokeh.models import LinearAxis, Range1d, SingleIntervalTicker
 from bokeh.plotting import Figure
-
-import colorcet as cc
 
 import numpy as np
 
@@ -127,137 +123,3 @@ def get_range(
     po_max = max(max(df.max()) for df in po_dfs_valid)
 
     return po_min, po_max
-
-
-class ColorSelector:
-    """
-    Class that provides API for color selection.
-
-    This class provides the API to get the color for each plot object
-    in the different rules.
-    """
-
-    @staticmethod
-    def create_instance(coloring_rule: str):
-        if coloring_rule == 'callback':
-            return ColorSelectorCallback()
-
-        if coloring_rule == 'callback_group':
-            return ColorSelectorCbg()
-
-        if coloring_rule == 'node':
-            return ColorSelectorNode()
-
-        if coloring_rule == 'communication':
-            return ColorSelectorComm()
-
-        if coloring_rule == 'publish_subscribe':
-            return ColorSelectorPubSub()
-
-    def __init__(self) -> None:
-        self._palette: Sequence[Color] = \
-            [self._from_rgb(*rgb) for rgb in cc.glasbey_bw_minc_20]
-        self._color_map: Dict[str, Color] = {}
-
-    def get_color(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        color_hash = self._get_color_hash(node_name,
-                                          cbg_name,
-                                          callback_name,
-                                          comm_name,
-                                          pub_sub_name)
-
-        if color_hash not in self._color_map:
-            color_index = len(self._color_map) % len(self._palette)
-            self._color_map[color_hash] = self._palette[color_index]
-
-        return self._color_map[color_hash]
-
-    @abstractmethod
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return
-
-    @staticmethod
-    def _from_rgb(r: float, g: float, b: float) -> Color:
-        r_ = int(r*255)
-        g_ = int(g*255)
-        b_ = int(b*255)
-        return RGB(r_, g_, b_)
-
-
-class ColorSelectorCallback(ColorSelector):
-
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return callback_name
-
-
-class ColorSelectorCbg(ColorSelector):
-
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return cbg_name
-
-
-class ColorSelectorNode(ColorSelector):
-
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return node_name
-
-
-class ColorSelectorComm(ColorSelector):
-
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return comm_name
-
-
-class ColorSelectorPubSub(ColorSelector):
-
-    def _get_color_hash(
-        self,
-        node_name: str,
-        cbg_name: str,
-        callback_name: str,
-        comm_name: Optional[str] = None,
-        pub_sub_name: Optional[str] = None
-    ) -> Color:
-        return pub_sub_name
